Add small scripts to generate random password
Will generate a password of the length specified by the user (or 15 if
no length given), then copy it to the clipboard (microsoft clipboard on
WSL, wl-copy if wayland, xclip if X11), then clear the clipboard 30
seconds later. mkpasswd_alnum.sh is the same, except it does not use any
special characters in case a website has finicky preferences.