Have neomutt use khard for address book
neomutt will sync addresses and emails now on 'S' with
vdirsyncer and mbsync, then if I want to add addresses,
can use 'A' to interactively go through each address in from, to, cc
and add it to the addressbook. I can use TAB to then add addresses when
composing mails.