Add pinentry-dmenu.conf
Use pinentry-dmenu on OpenBSD.
This prevents the issue of being on a different TTY than gpg-agent was
started, which will then lead to ssh and git (using ssh) commands
failing if the card needs to be unlocked and prompts for a PIN.

Any graphical pinentry would circumvent this, the dmenu one is just nice
and minimalistic.

You can also issue:
gpg-connect-agent updatestartuptty /bye

prior to an invocation of ssh and it would work still using the
tty pinentry.